Use a checksum of the secret to trigger re-deployment in case of change
With the current implementation, a change to the secret does not trigger a redeployment, which leads to pods using the old secret value. With this change, the checksum changes, will Helm will recognize as a change in the deployment, and re-deploy.

This logic is only applied if `global.secrets.useSecretName` is not set, since in that case, the secret is not part of the chart, and therefore cannot be used to calculate a checksum.